bowery-sms/requirements.txt
beautifulsoup4==4.3.2
Flask==0.10.1
gunicorn==19.3.0
httplib2==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-dateutil==2.4.2
pytz==2015.2
requests==2.6.0
six==1.9.0
smtpapi==0.2.0
twilio==4.0.0
Werkzeug==0.10.4
git+https://github.com/dgrtwo/ParsePy.git
